airflow/requirements.txt
apache-airflow==2.10.2
apache-airflow[statsd]
apache-airflow-providers-postgres
psycopg2-binary==2.9.9
python-dotenv
numpy==1.26.4
pandas
tqdm
dbt-core
dbt-postgres
astronomer-cosmos>=1.0.3 